% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildKernel.R
\name{buildKernel}
\alias{buildKernel}
\title{buildKernel}
\usage{
buildKernel(HDDDir = "~/git/linux", buildDir = "/tmp/linux",
  local = system(intern = TRUE, "uname -n"), rev = 4.11, jobs = max(1,
  parallel::detectCores() - 1), job = "bindeb-pkg")
}
\description{
buildKernel
}
