% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildKernel.R
\name{cloneKernel}
\alias{cloneKernel}
\title{cloneKernel}
\usage{
cloneKernel(
  kernelCloneUrl = "https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git",
  branch = "master",
  baseDir = "~/git"
)
}
\description{
cloneKernel
}
