% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildKernel.R
\name{buildKernel}
\alias{buildKernel}
\title{buildKernel}
\usage{
buildKernel(HDDDir = "~/git/linux", buildDir = "/tmp/linux",
  local = system(intern = TRUE, "uname -n"), jobs = max(2,
  parallel::detectCores()), target = "bindeb-pkg", install = TRUE,
  update = TRUE)
}
\description{
buildKernel
}
